Add transactions
Connections may now be tied to the current fiber, either for as long
as required (`Rome.checkout`) that *must* be released with
`Rome.release` or for the duration of a block (`Rome.with_connection`).

Transactions can be initialized with `tx = Rome.begin_transaction` and
must be terminated by calling `tx.commit` ot `tx.rollback`.
Unterminated connections will rollback automatically when
`Rome.release` is called.

Transactions can live the duration of a block using `Rome.transaction`,
which will tie the DB connection for the duration of the block
(unless already tied).

Transactions can be nested.